[ch60629] refactoring of the unit tests
